+++ Source/NonVisuals/Radios/RadioPanelPZ69FA18C.cs
@@ -1251,7 +1251,7 @@ namespace NonVisuals.Radios
                         PluginManager.DoEvent(
                             DCSFPProfile.SelectedProfile.Description,
                             HIDInstance,
-                            (int)PluginGamingPanelEnum.PZ69RadioPanel,
+                            PluginGamingPanelEnum.PZ69RadioPanel,
                             (int)radioPanelKnob.RadioPanelPZ69Knob,
                             radioPanelKnob.IsOn,
                             null);
